+++ src/blocks.js
@@ -11059,8 +11059,6 @@ MultiArgMorph.prototype.addInput = function (contents) {
     var i, name,
         newPart = this.labelPart(this.slotSpec),
         idx = this.children.length - 1;
-        // wantMore = false;
-    // newPart.alpha = this.alpha ? 1 : (1 - this.alpha) / 2;
     if (contents) {
         newPart.setContents(contents);
     } else if (this.elementSpec === '%scriptVars' ||
@@ -11080,14 +11078,9 @@ MultiArgMorph.prototype.addInput = function (contents) {
         if (this.is3ArgRingInHOF() && idx < 4) {
             newPart.setContents([
                 localize('item'),
-                localize('idx'),
-                localize('data')
+                localize('index'),
+                localize('list')
             ][idx - 1]);
-            /* commented out for now
-            if (idx === 1) { // fully expand to 3 items
-                wantMore = true;
-            }
-            */
         } else {
             newPart.setContents('#' + idx);
         }
@@ -11096,12 +11089,6 @@ MultiArgMorph.prototype.addInput = function (contents) {
     this.children.splice(idx, 0, newPart);
     newPart.drawNew();
     this.fixLayout();
-    /* commented out for now
-    if (wantMore) { // expand to 3 inputs in certain HOFs
-        this.addInput();
-        this.addInput();
-    }
-    */
 };
 
 MultiArgMorph.prototype.removeInput = function () {