+++ HISTORY.md
@@ -2,12 +2,14 @@
 
 ## in development:
 * **New Features:**
+    * expanding the rings in "map", "keep" and "find" shows 3 inputs named "item", "idx" and "data" 
 * **Notable Changes:**
 * **Notable Fixes:**
 * **Translation Updates:**
 
 ### 2019-11-01
 * new dev version
+* expanding the rings in "map", "keep" and "find" shows 3 inputs named "item", "idx" and "data"
 
 ## v5.2.5:
 * **Notable Fix:**